% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{app}
\alias{app}
\title{MarConsNetApp Conserved and Protected Area App}
\usage{
app()
}
\description{
This function creates a shiny app for the Maritimes region.
It addressed the following goals:
Site-Level Goals
- Report on what scientific work is occurring, and resources
allocated for this
- Report on how the scientific work being done is contributing to
indicators and therefore conservation objectives (CO)
- Report on the status of the sites, based on existing data
Network-level Goals
- How are individual sites contributing to network objectives
- How are science projects contributing to network objectives
- Provide scientific data to support status claims
}
\examples{
\dontrun{
app()
}
}
