% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getLatLon.R
\name{getLatLon}
\alias{getLatLon}
\title{Extract Latitude and Longitude from a Spatial Data Frame}
\usage{
getLatLon(sf = NULL)
}
\arguments{
\item{sf}{argument of class 'sf', 'tbl_df', 'tbl', or 'data.frame'.
Object must contain spatial data with a geometry column and either a
'NAME_E' or 'SiteName_E' column.}
}
\description{
This function extracts the latitude and longitude coordinates from a spatial
data frame to display protected area polygons on the map in the app.
}
