% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NAME_to_tag.R
\name{NAME_to_tag}
\alias{NAME_to_tag}
\title{Change names from NAME_E to PPT Tags}
\usage{
NAME_to_tag(df = NULL, names = NULL)
}
\arguments{
\item{df}{a data frame with names "flower", "place", "tab", "link"}

\item{names}{a character string to clean up}
}
\description{
This function extracts the names of the protected and conserved areas
 from `data_CPCAD_areas` and converts them to tags used in the
Project Planning Tool
}
\details{
Place holder
}
\examples{
\dontrun{
# FIXME
}
}
