% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_geom_data.R
\name{process_geom_data}
\alias{process_geom_data}
\title{Process Geometric Data}
\usage{
process_geom_data(x)
}
\arguments{
\item{x}{an sf object or data frame containing geometric data}
}
\value{
an sf object with 'year' and 'geometry' columns
}
\description{
Process Geometric Data
}
