% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_plots.R
\name{save_plots}
\alias{save_plots}
\title{Save plots}
\usage{
save_plots(df, STORE = dirname(path_to_store()))
}
\arguments{
\item{df}{one of the objective dataframes. Must contain columns: data, plot, areaID, indicator}

\item{STORE}{path to store (default as the dirname(path_to_store()))}
}
\value{
character vector of filenames where plots were saved
}
\description{
Save plots
}
\examples{
\dontrun{
# Example data frame
df <- data.frame(
 data = I(list(mtcars, mtcars)),
 plot = I(list(ggplot2::qplot(mpg, wt, data = mtcars), ggplot2::qplot(mpg, wt, data = mtcars))),
 areaID = c("Area1", "Area2"),
 indicator = c("Indicator1", "Indicator2")
 )
 save_plots(df)
 }
}
