% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getLatLon.R
\name{getLatLon}
\alias{getLatLon}
\title{Extract Latitude and Longitude from a Spatial Data Frame}
\usage{
getLatLon(sf = NULL)
}
\description{
This function extracts the latitude and longitude coordinates from a spatial data frame to display
 protected area polygons on the map in the app.
}
\details{
@param sf argument of class 'sf', 'tbl_df', 'tbl', or 'data.frame'. Object must contain spatial data with
 a geometry column and either a 'NAME_E' or 'SiteName_E' column.
 @importFrom sf st_sfc st_coordinates sfc_multipolygon

 @export

 @examples
   \dontrun{
   polygons <- getLatLon(sf_data)
   print(polygons)
   }

 @return A list where each element corresponds to a polygon in the input 'sf' object. Each element
 is a list with the following components:
 \describe{
   \item{lat}{A vector of latitude coordinates.}
   \item{lng}{A vector of longitude coordinates.}
   \item{color}{the color assigned to the polygon, which is "hotpink".}
   }
}
