% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path_to_store.R
\name{path_to_store}
\alias{path_to_store}
\title{path_to_store}
\usage{
path_to_store()
}
\value{
character
}
\description{
Convenience function that returns our commonly used paths to the targets data store
}
\examples{
STORE <- path_to_store()
}
